My previous work on Gui Main Menu (Title Screen)
This is my previous work implementation of the Gui Main Menu (Title Screen). I name this file MainDev.py a separate file so that it can be executed to see the execution Main Menu(Title Screen) display.
I created a class called MainMenu that is a subclass of the sgc.Menu class. 
This is the class with which the main menu was implemented. To implement the main menu,
 I modified the main function by moving most of its contents to a new function I named new_game().
 I also copied some of its contents over to the open_replay() function. The main function was then dedicated to processing the user's selection in the main menu.